feat: make service agnostic on repo impl
