+++ src/main/resources/application.properties
@@ -5,7 +5,7 @@ quarkus.swagger-ui.always-include=true
 quarkus.native.additional-build-args=-march=compatibility
 quarkus.keycloak.devservices.realm-path=quarkus-realm.json
 quarkus.keycloak.devservices.port=32794
-quarkus.wiremock.devservices.port=4000
+quarkus.wiremock.devservices.port=4001
 quarkus.oidc.client-id=backend-service
 quarkus.oidc.credentials.secret=secret
 quarkus.openapi-generator.codegen.spec.discovery_yaml.enable-security-generation=false
@@ -27,13 +27,14 @@ quarkus.openapi-generator.codegen.spec.keycloak_yaml.enable-security-generation=
 quarkus.openapi-generator.codegen.spec.keycloak_yaml.base-package=io.github.genomicdatainfrastructure.discovery.remote.keycloak
 quarkus.openapi-generator.codegen.spec.keycloak_yaml.additional-model-type-annotations=@lombok.Data;@lombok.NoArgsConstructor;@lombok.AllArgsConstructor;@lombok.Builder;@io.quarkus.runtime.annotations.RegisterForReflection
 quarkus.openapi-generator.codegen.spec.keycloak_yaml.generate-part-filename=false
-quarkus.rest-client.ckan_yaml.url=http://localhost:4000
-quarkus.rest-client.keycloak_yaml.url=http://localhost:4000
+quarkus.rest-client.ckan_yaml.url=http://localhost:4001
+quarkus.rest-client.keycloak_yaml.url=http://localhost:4001
 quarkus.rest-client.keycloak_yaml.beacon_idp_alias=LSAAI
-quarkus.rest-client.beacon_yaml.url=http://localhost:4000
+quarkus.rest-client.beacon_yaml.url=http://localhost:4001
 quarkus.rest-client.beacon_yaml.read-timeout=60000
-%dev.quarkus.oidc.auth-server-url=https://keycloak-test.healthdata.nl/realms/ckan
-%dev.quarkus.oidc.client-id=ckan
+%dev.quarkus.oidc.auth-server-url=https://id.portal.dev.gdi.lu/realms/gdi
+%dev.quarkus.oidc.client-id=gdi
+%dev.quarkus.oidc.credentials.secret=vjKOWL7bU3hzESRegDpprHjPkHoyoMAO
 %dev.quarkus.rest-client.ckan_yaml.url=https://ckan-test.healthdata.nl/
 %dev.quarkus.rest-client.keycloak_yaml.url=https://keycloak-test.healthdata.nl/realms/ckan
 %dev.quarkus.rest-client.beacon_yaml.url=https://beacon-network-backend-demo.ega-archive.org/beacon-network
@@ -42,5 +43,9 @@ quarkus.rest-client.beacon_yaml.read-timeout=60000
 %dev.quarkus.log.category."org.jboss.resteasy.reactive.client.logging".level=DEBUG
 quarkus.http.cors=true
 quarkus.http.cors.origins=http://localhost:3000
+quarkus.http.port=8082
 quarkus.http.cors.headers=accept, authorization, content-type, x-requested-with
 quarkus.http.cors.methods=POST, OPTIONS, GET
+
+repositories.gdi=true
+repositories.lndc=false
